requirements.txt
pandas
numpy
yfinance
requests
scikit-learn
joblib
python-dotenv
flask
dash
plotly
